src/requirements.txt
numpy~=2.2.6
pyyaml~=6.0.2
scipy~=1.16.1
opencv-python~=4.12.0.88
opencv-contrib-python
librosa~=0.11.0
dash~=3.2.0
matplotlib~=3.10.5
plotly~=6.3.0
ultralytics~=8.3.179
scikit-learn~=1.7.1
Flask~=3.1.1